Elegantly handle "past" timeouts with a "past strategy"
Each timeout will now be registered with a "past strategy", which indicates
how it should be treated if it's registered for sometime in the past.
Some timeouts need to be deleted, punted, or invoked immediately.
The "Noon" timeout is now registered in series with the "Pre-Noon" timeout
rather than in parallel, which will make the sequencing more resilient.