Add "begin home stretch" event
When the season completion reaches 90%, a daily event will happen
that triggers a variety of "surprises". This includes doling out point
multipliers, which this commit also adds support for. The event requires
a new "recursive" sort of timeout which grabs "surprise" types from a
queue until it's exhausted. This commit also refactors some canvas logic.