Finalize submissions early if all votes are in
The reveal of the results can occur at either of the voting reminder times
if all votes have been cast at this point. There are stop-gaps in place to
prevent the finalizing from happening again once the vote data is wiped.
This makes me think that the way submissions state data is stored should
get a refactor soon...