Only send the magic word hint below a top score percentage
Originally, the hint was sent to anyone below the top 3 players.
This season, the top ~5 players are all within a couple points of each
other, so sending to player 4 or 5 would be totally unfair. This change
makes it so that only players below 85% of the top score can receive it.
Also adds several interesting custom words.