Break vote ties using secondary "prioritize golds" strategy
If the scores of two submissions are the same, we currently break ties using
total number of votes (e.g. 3 bronzes > 1 gold), but this doesn't handle scenarios
with equal number of votes (e.g. 2 silvers = 1 gold + 1 bronze). Break this tie
using the secondary strategy of priotizing number of gold votes.