Determine the leader nerf threshold when the day starts
Previously, this value was more "dynamic" because it was just determined on the
spot at 90% of the top score. This meant that the order in which players said GM
could really affect who would get nerfed. It seems more fair to set this as a
predetermined static value on each day.