Make the nerf threshold a constant amount below the top score
Previously, the threshold was some percentage of the top score, which made the range way
too large near the end of the season. Now, it'll be 2 top awards below the top score rather
than 90% of the top score. To account for the increased range in the early season, the nerf
will apply 20% into the season rather than 10% in.