Add dungeon item support
The trap action and the new boulder action now require items rather than points.
Items are awarded to users via the abstract game's "award major" and "award minor" APIs.
The major prize is for first-place contest winners, and the minor prize is TBD.
The boulder item is used to place permanent walls, which can currently be abused.
The follow-up is to add logic that will detect if a boulder softlocks players.