Use game ranking to order players in general
Rather than always looking at how many points a user has that week, look at their ranking in-game.
This requires us to maintain a rank field for each player in the dungeon and refresh it often.
This will allow the magic word recipient to be selected based on their position in-game.
Also fixes the core game logic to create a sectional dungeon rather than a central one.