Prevent players from softlocking the game
If a player tries to place a boulder or seal a doorway in a way that will prevent players
from reaching the goal, knock them out and end their turn. Also added decision validation
for the seal action. Even with decision validation, this can still happen if two players
use boulders/seals in the same turn. This change also refactors adjacent location logic.