Collapse "bump and gave up" statements
If multiple consecutive players bump into someone and give up (which will happen a lot more
now that we do decision-shuffled ordering), then collapse the statements into one statement
which will be easier to read. Also, for non-goner bumps, add the name of the blocking user
once again. Also, only continue processing immediately if the one statement is a step.