Introduce voting probation system
For anonymous submissions, if a player fails to vote then they'll be placed on probation until
the next time they successfully vote. When on probation, the bot will not wait for their vote,
meaning that the submission voting will be finalized once everyone who is not on probation has voted.