Support monthly "high-effort" submission prompts
On the first Saturday of the month, the sungazers channel will be polled for the next submissions prompt.
The rest of the players will be notified of this prompt at each pre-noon until the event starts.
Unlike standard prompt polling, this variant will retry if there are no suggested prompts.