Use scaled points distribution for awards
Rather than the previous hardcoded points-by-rank system.
This way, the distribution is more dynamic and makes a difference regardless
of whether a player is near the front or back, whereas the previous system
only really made a difference for those in the top few spots. In other words,
rank is more relative (e.g. being 10th place out of 100 players can still
get you a good amount of points)