Award prize for pre-game submission victory
Since the game isn't initialized until the first decision of the season,
the first contest can't directly award a prize to the winner. This commit
tracks the last contest's winner directly in the state and awards the first
week's winner at the time of game creation.

Updates Classic to use endTurn for computing winners, and updates Maze to
use endTurn for wiping item offers (since the first week's offer would be
useless if awarded and wiped on Saturday morning).