Pick out fallback users for Popcorn
If the user tags himself or fails to tag someone else, pick a "fallback" for them.
Picks the most active user who hasn't gone yet. The idea here is that (1) less popular
players may need help being picked, and (2) it's annoying when you forget to tag someone
and the bot treats it as a mistake rather than just rolling with it.