Validate that selected users are in the game
For the generic user select game decision component, players who aren't
in the game (e.g. the bot itself) will show up. There's no way to control
this AFAIK, so validate that the selected player is in the game. Otherwise,
the user ID will be passed off to the game logic and strange behavior may occur.