Use OpenAI to summarize the Popcorn story
At noon, all the messages comprising the Popcorn story will be compacted into one
large payload and sent to OpenAI, where it will be compressed into a 300-word
synopsis and sent to the GM channel. This probably needs to be cleaned up and
improved, but I just wanna get it working for now.