Rework Risk attacker/defender logic
Refactors the conflict state representation to have any number of attackers
and one optional defender. This more generic representation allows for a
multi-pronged, round-robin attack against one single territory. When there
is no defender, all the attackers fight in a circular fashion.

Next step is to fully test the circular conflict logic and improve the renders.