Try to handle MP2 non-start auctions gracefully
A weird bug can happen where the GM timeout can fail halfway through, beginning the game
turn yet never invoking any of the decision phases. In this case, don't process the auctions
for that day since they were never actually introduced. Also, refund the forcer's force item
in the case of a non-start. Try to "fix" this bug by using the safe messenger utility when
sending the render of the game state (I think that's what caused this).