Move MP2 auction dates to later in the morning
First auction is at mid-morning, second is at three-quarters-morning, and so on...
This is to handle any number of auctions in a turn, but also to make the bidding
more frantic near the end of the morning when people are paying attention.
As part of this, support was added for "relative" decision phase times.