Fix Candyland space rendering, add new assets
Finally found a proper solution for the fatal error that would occur whenever
trying to alter the pixels in the board image data. This removes the previous
workaround (using a pre-rendered board), splits each space into a separate
image and uses them as a mask (similar to Risk rendering). Adds assets for
shiny cards, logic for rendering negative cards, and assets for
dunce/rainbow/black cards.